// types/enums.ts
export enum UserType {
  ADMIN = "ADMIN",
  DEFAULT = "DEFAULT",
  CADASTRE = "CADASTRE",
  REPORT = "REPORT",
}

export enum UnitType {
  KG = "KG",
  G = "G",
  L = "L",
  UN = "UN",
}

export enum ProductType {
  DONATED = "DONATED",
  PURCHASED = "PURCHASED",
}

export enum ReportType {
  VALIDITY = "VALIDITY",
  DONATIONS = "DONATIONS",
  PURCHASED = "PURCHASED",
  INVENTORY = "INVENTORY",
}

export enum PdfUnitType {
  PT = "pt",
  MM = "mm",
  CM = "cm",
  IN = "in",
  A4 = "a4",
}